REWORKED THE WHOLE THING
For now doesn't even use time as the x-axis, as I found that some points seemed to be separated by a huge margin, resulting in the interesting ones being squished into a single column. So for now I'm just using longitude as my x-coord until we decide what to do. (to make something at least visible.)