This is a pretty major (and yes, ugly) change. Basically I got the "distance" view working with the selection/moving points etc. More or less anyway. (Jill made it clear to me that everyone was mad at me for breaking it so hopefully we're all happy again.)
Anyway. I'm sure there are lots of bugs and broken things and unnecessary things I did. If you see something you don't like, just be brave and change it for the better. I wish I had more time to spend on this but I absolutely have to work on other commitments now.

Unfortunately, I think it runs a bit slower now that I update this distance information every refresh. If anyone wants to add some sort of "dirty" flag to mark whenever a point is moved so this doesn't need to be done so frequently, that might help.